pyproject.toml
[project]
name = "flight-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.1.1",
    "google-generativeai>=0.5.4",
    "pydantic>=2.11.7",
    "sqlalchemy>=2.0.43",
]
